an improvement of the Bellman–Ford algorithm
Given a weighted directed graph {\displaystyle G=(V,E)}{\displaystyle G=(V,E)} and a source vertex {\displaystyle s}s, the SPFA algorithm finds the shortest path from {\displaystyle s}s, to each vertex {\displaystyle v}v, in the graph. The length of the shortest path from {\displaystyle s}s to {\displaystyle v}v is stored in {\displaystyle d(v)}{\displaystyle d(v)} for each vertex {\displaystyle v}v.

The basic idea of SPFA is the same as the Bellman-Ford algorithm in that each vertex is used as a candidate to relax its adjacent vertices. The improvement over the latter is that instead of trying all vertices blindly, SPFA maintains a queue of candidate vertices and adds a vertex to the queue only if that vertex is relaxed. This process repeats until no more vertex can be relaxed.

Below is the pseudo-code of the algorithm.[4] Here {\displaystyle Q}Q is a first-in, first-out queue of candidate vertices, and {\displaystyle w(u,v)}{\displaystyle w(u,v)} is the edge weight of {\displaystyle (u,v)}(u,v).